Refactoring playable items
