+++ spotlight/commands/search_songs.py
@@ -1,5 +1,8 @@
 from caching.holder import CacheHolder
-from spotlight.suggestions.play import SongSuggestion, QueueSuggestion, PlaylistSuggestion, AlbumSuggestion, ArtistSuggestion
+from spotlight.suggestions.playable.album import AlbumSuggestion
+from spotlight.suggestions.playable.artist import ArtistSuggestion
+from spotlight.suggestions.playable.playlist import PlaylistSuggestion
+from spotlight.suggestions.playable.song import SongSuggestion, QueueSuggestion
 from spotlight.suggestions.templates import FillSuggestion, PassiveSuggestion
 from spotlight.commands.command import Command
